Part 4 exercise 4.21 done

+++ osa4/blogilista_backend/controllers/blogs.js
@@ -39,8 +39,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  //checking that token field in request matches decodedToken with envSecret value
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  //checking if tokenid and blog's userid match
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  }
+
+  return response.status(401).json({ error: 'delete unauthorized for user' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
